Fix OverlayNG floating noder problem with repeated points
JTS' commit dr-jts@b9541cf3843d6a6e88192aaf24531ad304a42e1c
